packages/modeldb-interface: move Effect to src/types.ts
